move custom args into a folder
