+++ fast64_internal/sm64/settings/properties.py
@@ -18,7 +18,7 @@ from ...render_settings import on_update_render_settings
 from ...utility import directory_path_checks, directory_ui_warnings, prop_split, set_prop_if_in_data, upgrade_old_prop
 from ..sm64_constants import defaultExtendSegment4
 from ..sm64_objects import SM64_CombinedObjectProperties
-from ..sm64_custom_cmd import SM64_CustomCmdProperties, draw_custom_cmd_presets
+from ..custom_cmd.properties import SM64_CustomCmdProperties, draw_custom_cmd_presets
 from ..sm64_utility import export_rom_ui_warnings, import_rom_ui_warnings
 from ..tools import SM64_AddrConvProperties
