+++ fast64_internal/sm64/sm64_collision.py
@@ -10,7 +10,7 @@ from .sm64_constants import (
 )
 from .sm64_utility import export_rom_checks
 from .sm64_objects import SM64_Area, start_process_sm64_objects
-from .sm64_custom_cmd import CustomCmd
+from .custom_cmd.exporting import CustomCmd
 from .sm64_level_parser import parseLevelAtPointer
 from .sm64_rom_tweaks import ExtendBank0x04
 from ..panels import SM64_Panel
